squash! Revert "Use sigaction(SA_NOCLDWAIT) for SIGCHLD handling"
;240623 not actually reverted.  use make HAVE_USE_SIGACTION_SIGCHLD=1
to follow the new code path.

my wait(3) says

	"As specified in Consequences of Process Termination, if the
	calling process has SA_NOCLDWAIT set or has SIGCHLD set to
	SIG_IGN, then the termination of a child process will not
	cause status information to become available to a thread
	blocked in wait(), waitid(), or waitpid().  Thus, a thread
	blocked in one of the wait functions will remain blocked
	unless some other condition causes the thread to resume
	execution (such as an [ECHILD] failure due to no remaining
	children in the set of waited-for children)."

so commit 712d6639ff8e seems to make it impossible to guarantee that
dwm can do a fork + exec + wait for child to exit + read the output

the pattern of calling dmenu: to fork, exec, call wait(NULL) on the
parent, and then reading the child's response from the pipe now hangs
forever.